refactor(api): create the AssessmentDtoFactory to break circular dependency
