+++ api/src/shared/application/assessments/assessment-controller.js
@@ -1,3 +1,4 @@
+import { AssessmentDtoFactory } from '../../domain/models/AssessmentDtoFactory.js';
 import { sharedUsecases } from '../../domain/usecases/index.js';
 import * as assessmentSerializer from '../../infrastructure/serializers/jsonapi/assessment-serializer.js';
 import { extractUserIdFromRequest, getChallengeLocale } from '../../infrastructure/utils/request-response-utils.js';
@@ -17,7 +18,8 @@ const getAssessmentWithNextChallenge = async function (
     locale,
   });
 
-  return dependencies.assessmentSerializer.serialize(assessment.toDto(globalProgression));
+  const assessmentDto = AssessmentDtoFactory.toDto(assessment, globalProgression);
+  return dependencies.assessmentSerializer.serialize(assessmentDto);
 };
 
 const assessmentController = {
